Add sensoranalytics logrus prereq, mapping-agent url
In sensoranalytics the Logrus Go package is required, so add that to the
Prerequisites. Also make Go package links consistent by using the godoc.org
addresses for each.

In the mapping-agent development section of that demo, the URL you need to go
to needs to have static/ in it, so add that as well.

Both completed for both 1.8 and 1.9.