Make 1.9 updates and improvements to fastdata-iot
Add a link to the code repository where the Docker Hub images get their code.

Update all references to 1.8 to be 1.9.

Add import_data.sh file to a new bus-demo directory from the Docker image and a
related README.md so it's more clear what we're running to configure Cassandra.

Update inline JSON and Spark job command to use the Cassandra vip.

Update akka-ingest and dashboard .json files so they're the same as what's in
the README.md.

Update JSON screenshot so it matches our code, and new Marathon screenshot.